Remove files not required for the graph
handlers are the rules for a specific MUD and belong in the MUD, not the
graph or the MUD engine, although the MUD engine could have some
standard rules, like logging in.

Attack and defence modules definitely have nothing to do with the graph
as they pertain very clearly to a game, not a general purpose rules
engine.

conn and conn_sup are for connecting to a MUD.

test socket is for connecting to a MUD, nothing connects to the graph
with anything like a socket. Same goes for web sockets.

gerlshmud_world is for a MUD world.

search, roll and parse are all MUD-related.