Remove unnecessary GERLSHMUD files
The graph won't do the logging or know about "handlers" (rules modules)
or web sockets, or ... I don't think ... how to detect the protocol.

I'll have tests for EGRE, but I'll start from scratch.

EGRE will be a dependency in an app and doesn't need the existing config
in sys.config or the vm.args.

Scripts to run an app using EGRE will come from the top-level app. EGRE
will be a library that's included.