Add jsx library for logging
I've heard that jiffy is faster, but I don't care.
Since this is a proof-of-concept project I try to work off the
assumption that I have infinite CPU, infinite memory and infinite
bandwidth. i.e. "Here's a neat idea, I wonder if it could *possibly*
work", not "Is this a _good_ idea."