Test attempt
The first step in propagating a message through the rules graph is
firing off a message "attempt" at an object in the graph.

Objects have handlers. Each handler can judge whether or not a message
succeeds or fails, or the object can obstain from judgement if it's not
interested in that message. Objects "abstain" by returning 'undefined'
from the attempt/1 function.

Any object can, and probably will, have multiple handlers to handle
massaging messages and communicating with other handlers on this and
other objects; the first handler for an object that passes judgement on
the message short circuits the judgement for all handlers on that
object: as soon as a handler returns succeed or fail, no other handlers
on that object are called. Handlers not interested in a message should
abstain from judging the success or failure of the message in order to
let other handlers for that object "see" the event.

I'm building this rules engine for text-based games and there are two
different kinds of actions that happen in a MUD:
- commands the player types that need to be parsed and checked for
  sensibility.
- triggers that happen in response to commands.

For example, typing "pick lock" is different from an event "player is
picking the lock" that might be triggered from that command.

Basically, "is this a thing that can happen?" and then "what does this
action trigger, if anything?"