Add attempt_after test, test auto-sub
There's two ways that an object can subscribe to the result of a
message:
1) when calling egre_attempt/2,3
2) when a handler module handles the attempt

I discovered while attempting to test egre_object:attempt_after/2
that calling attempt with no ShouldSub argument, or with ShouldSub =
true, the list of subscribers gets set to [self()]. When we call
egre_object:attempt/2,3 from the common test process, self() is the test
process; when we call egre_object:attempt_after/2, self() ends up being
the object process: this is because attempt_after first sends a message
to the object process telling it to call attempt/2, and since it calls
attempt/2 _itself_, then self() is the object.

So, since there's two ways of subscribing, I split the attempt test into
two: one that subscribes off the bat, and one that doesn't. In order to
achieve this I caused both of the tests to use attempt_after so that
attempt/2,3 would be called from the object process itself, so that
self() is the right process, whether subscribing or not. That way, if I
specify not to subscribe, I can make sure that succeed/1 doesn't get
called within the attempt test handler. Previously, when the attempt
test accidentally subscribed off the bat, the common test process
actually got the succeed message. So, I thought I was making sure that
succeed messages weren't being sent, but they were being sent to the
wrong process. In order to support this I added sub/nosub support to the
attempt_after function. Also, I decided that handle_info({Pid, Msg}) was
a little too obscure, so I changed it to handle_info({send_after, ...}).
This will hopefully clarify that the only handle_info/2 call getting a
pid and a message is from the send_after/2,3 function; the
handle_cast_/2 function makes it clear where a message comes from with
the {attempt, Msg, Procs} tuple.

I also added the attempt_after test to make sure that the message really
is fired after a delay.