Add tests: set, populate, second order sub
Test set/2 - set a property on a running object

Test populate - send a running object a list of ID+Pid pairs and have it
replace every property value that matches an ID with the corresponding
Pid.

For instance, if object 1 has properties [{obj2, id2}], and populate is
called for object 1 with [{id2, SomePid}], then object 1's properties
will be [{obj2, SomePid}]; that is, the actual ID 'id2' (must be an atom
or it won't work) is replaced with SomePid. Not that 'obj2' is what
object 1 refers to the other object as. This mechanism is so that you
can set up a MUD world with objects referring to each other by ID in the
properties, and then once all the object sare created and have Pids, you
can go back and tell them all to update their refernce-by-ID values with
the Pids matching the IDs. Simple, no? lol

Test "2nd order sub" - test that subscriptions work with a simple chain
of objects: object 1 gets a message and ignores it; however, object 1
has a reference to object 2 in its properties, so object 2 also gets the
message and then succeeds or fails and subs or doesn't sub. I just want
to make sure the messages are actually travelling through the graph and
subscriptions are being recorded.

I don't know of a simple function for random atoms so I made one.