Update README, change to README.md
I think the reason my mermaid sequence diagram isn't showing up is that
GitHub expects triple-backtick-codeblocks to be in .md files.

Changed "event" to "message" with an explanation in the README itself.

Added a couple paragraphs about my initial motivation for the MUD that
EGRE came out of.

Updated my mermaid sequence diagram to, hopefully, make it a little
clearer.