Fix & test broadcasting
Don't broadcast to self!
I wouldn't have picked up on this if I didn't write this test: turns out
the broadcast just kept looping and looping.

Broadcast is used to set a property on an entire sub-graph: e.g. if an
entire sub-graph is moved from one object to another, sometimes we'll
need to tell every object in the sub-graph what object it's been moved
to, e.g. 'character'.

Now that I think about it, I'm not sure I needed to convert the
broadcast attempt call to a attempt_after: I did this before I discovered that
the 'pid' property (i.e. self) was being included as one of the objects
to broadcast to; my logic was that I don't want the broadcasting object
to automatically subscribe to the message being broadcast: broadcast
called attempt directly, which adds self() to the list of suscribing
objects: that means the _broadcaster_ was automatically subscribing to
the broadcasted message, rather than the reciptient of the message
subscribing to it.
By using attempt_after/2, the recipient object gets the 'send_after'
message and then calls attempt/2 itself. Since the recipient object
calls attempt/2, it will auto-subscribe itself to the event. That is,
the recipient object PID will be returned by self().

Should objects auto-subscribe to broadcast messages? That's a question
for another day.