Test {resend, TargetPid, NewMessage} result
The way to communicate between processes is by sending messages that
essentially have blanks to fill in; you decide what the "question" or
the blank will be and then when a target object knows how to fill in the
blank (e.g. with its own pid) then that object updates the message and
resends it.

So coordination is done by resending the message out on the graph over
and over until either all the questions are answered or the message
succeeds by default with, probably, no object subscribed to the result.

The resend result resets the #procs{} record, which tracks
subscriptions, so it doesn't matter what value the object returns for
the subscription.

In this test I'm abusing the ability of attempt/1 message handlers to
update the objects props. In reality, or at least in the context of a
MUD, an attempt handler probably shouldn't update the properties of the
object since there's no guarantee that the message actuals produces an
affect.