Objects provide pid filters, re-test broadcast
When the graph code needs to know where to send a broadcast, a stop, or
a messaage, the object will provide a property filter to find the pids
to send to, unless the default of {_key, Pid} when is_pid(Pid) works
fine.

This means that the graph will allow a user of the library to specify
what properties contain pids that should be broadcast too, which
properties that contain pids that should be stopped, and which
properties contain pids to propagate regular graph messages to.

I had screwed up the broadcast test previously anyway, so I fixed it:
previously I was assuming that a broadcast was multi-level: it's not. If
you want a broadcast to go past one level of objects, objects that are
broadcast to have to re-broadcast the message. That is "broadcast"
doesn't mean "send to a sub-graph", or "send to all children", it just
means "send to all connected objects at once" instead of just the
next _one_ connected object. It's what the name implies; however, I
started to confuse what "broadcast" means, with what I was using it for;
i.e. to pass a message to an entire sub-graph: e.g. all the body parts
on a character.