DO NOT AUTO SUBSCRIBE EGREMUD_CONN
Add an API attempt call that allows the caller to opt out of subscribing
to the result of the event being attempted.

We ran into a bug where the egremud_conn object was sending a message
into the graph and was auto-subscribed to the result. egremud_conn is
not a graph object and didn't know what to do with {succeed, {...}}
message that it got for the event that it sent into the graph.

There are three "connection" objects:
WebSocket/test socket -> egremud_conn -> graph connection object
Only the graph connection object should participate in graph events.

We also left in a couple commented out ct:pal calls which proved very
handy for seeing what processes were subscribing to what events. It
didn't show us how egremud_conn was subscribing, but it showed us that
it *was* subscribing, which made no sense, it not being a graph object.