Remove MUD-specific code
EGRE shouldn't know anything about the application using it, so all the
MUD stuff should be removed.

EGRE now relies on an "extract record fun" and a "serialize" fun.

The extract record fun will take the properties of a graph object and
extract a record to be passed to rules modules, and a list of properties
to be logged. To day, the list of properties were the "parents" of the
object, e.g. owner, body_part, top_item, and character. EGRE_MUD_1 now
sets EGRE application env vars to supply those functions.

The serialize fun is for converting values to binary for exporting as
JSON. EGRE shouldn't know anything about a #body_part{} or a
 #top_item{}. Since EGRE_MUD_1 still needs to serialize values that
aren't specific to itself, EGRE passes a serializer that EGRE_MUD_1
could use. This could create quite the stack if the two applications
keep calling each other back and forth, but I don't see that happening.