Multiple logging channels, DB logging
In order to support DB logging I separated json logging out of the
egre_event_log and added both egre_event_log_json and
egre_event_log_postgres. Now log messages go both to a file, in JSON,
which can be wrapped in JavaScript and loaded in a web page, and to a
Postgres database. I'm using the epgsql Erlang Postgres driver with the
epgsqla module: I could have gone the easy route and used epgsql, but
somewhere in the README it referred to that module as "old" and
"backward-compatible", so I opted to figure out the asynchronous
version. The DB logging is "working": there is more work to get useful
log messages out of the event properties.

The reason for the Postgres DB logging is so I can query for various
logging events by time, by event, by rules handler module, etc. The
structured logging page is great, but I'd like to do more queries:
instead of writing my own query-able front-end on the logging web page,
it seemed wiser to just use a database.

I've added a postgres module, egre_postgres. So far, it only handles
one query, which is for inserting log properties into the "log" table.
egre_postgres runs as a gen_server and holds a link to the epgsql
gen_server, which holds the connection to the DB. egre_postgres is
started by the egre supervisor.

Since both the JSON and DB logging need serialization, I created a new
egre_serialize module. The module doesn't serialize to binary, only to
JSON-compatible data types, so the egre_event_log_postgres code has to
do some conversions to binary once the data is "serialized" (i.e.
"made JSON-friendly").