Fix parse transform attempt catch-all clause
I already had a catch-all case for the attempt, but it was limited to a
couple specific variables. I've changed it so that if the whole attempt
(Parents, Props, Event) is captured by a single variable, then we just
ignore it. I don't think I have any cases where it would make sent *not*
to match inside the attempt argument tuple to access at least the props
and the event.