Use gen_server propertly for Postgres logging
Move all the receive calls when interacting with epgsqla to
handle_info/2 calls instead of handling them manually. That's exactly
what the gen_* modules are for.

Insert binary spaces in between tuple elements so that when they are
serialized, they're not smushed together in the DB.

Update the DB log table to have more columns.