Uses gen_statem for DB logging
Because we're using epgsqla, everything we do with the database is
asynchronous: by using a state machine, we can step through each stage
of starting up the database. By using gen_statem, we can use the
"postpone" action to selectively receive only the messages we're
interested in until we're ready.

Co-authored-by: Terry Wilcox <terry.wilcox@mac.com>