Keep "wait for DB" in one state
It seemed easier and cleaner to keep the "wait for DB" timeout within
the 'logging' state in the egre_posgres gen_statem.

The "wait for DB" functionality is to be signaled that a period of
inactivity has happened, which implies that all logs from tests have
been inserted into the database.

The gen_statem documentation is pretty confusing, but you can have a
{timeout, Millis, Message} action as the last parameter of your
gen_statem state callback function. This is an "event timeout": if
another event comes in before the timeout, the timeout is canceled.

In order to keep the timer going we wrapped the 'logging' state
functionality in an additional function: the timer code is handled in
the outer 'logging' state, and the internal logic is handled in the
inner 'logging_' function.

Co-authored-by: Terry Wilcox <terry.wilcox@mac.com>