Make Postgres logging nicer and more compact
Add "Sub" if a module subs to an attempt; uses the 'subscribe' column.
Shrink attempt, succeed, and fail down to A, S, and F; put them in their
own columns.
Convert undefined to an empty string so it doesn't clutter up rows.
Change 'no_rules_module' to an underscore so it's not as visually
cluttered.
Remove 'rules_' from rules modules because it's redundant and wastes
space.