Attempt/1 returns #return{}
All succeed calls now take a context which represents the state of the
world at the time the event was being processed.

The #result{} is a more convenient way to allow attempt/1 clauses to
only return some of the possible return values. It also standardizes the
return values which can prevent errors: i.e. a touch of static typing
but only a constructor and no methods.