Subsequent DB timeout waits were too long
First off, the "wait for DB" logic needs to keep track of how much time
has been spent waiting so we don't rely on the gen_statem:call/3
timeout.

This updates the "wait for DB" code to first wait up to the specified
millis, but if the timeout is interrupted by another event coming from
the process mailbox, only add 30 more milliseconds. This means that if
we set a timeout for 500 ms and then immediately get another event,
we'll only wait a further 30 ms before we consider the DB to have
"quieted down" and we can stop the tests. It would be nice if the
gen_statem timeout code told you how much of the timeout was left.