Don't wait for ever for DB to be done
Decrement the amount of time we set in timeout timers when we're waiting
for the DB to finish writing but we get straggler log events.

Previously I kept resetting the timer to 50 ms or something when I got a
log event while waiting stop the DB.

The gen_statem:call/x timeout is now 10% more than the DB event wait
timeout amount. i.e. if you say to wait 500 ms for the DB to be done
writing the remaining log messages, then you wait 550 ms before you
timeout the gen_statem:call/x.