Allow {stop, _, _, _} to be returned from fail/1
I had a case where a failure of an attempt should shut down a process.

Also, if a process rules handler returns {stop, Reason, _, _}, then that
Reason should be the shut down reason.

Add another populate scenario where a world property has:
{Key, {{pid, Value1}, {pid, Value2}}}.
i.e. a Key that points to a two-tuple where both the values in the
two-tuple are each IDs that need to be converted to PIDs. In EGRE_MUD_1
I have a property called 'reserve' that points to an escrow process and
a PID of a process that needs to be reserved. At this point I might just
need to check every part of the world data structure and convert it to a
tuple.