Comment about not using the index for sub-props
See previous commit. Right now looking up the ID and icon of each
sub-property is redundant because the DB and web page already build an
index of PID -> ID from the top level log messages. It kind of makes
sense to use the index for that, so it's done in one place, but I think
it might be super slow: there's a TON of Mnesia calls when I run cprof.