Parent/prototype PID properties mask real PIDs
When an effect process is created from a prototype, the prototype
objects {pid, PID} property is still in the properties. I don't know
exactly why, but this masks the new {pid, PID} property that EGRE adds
to every new object. The solution is to make sure any existing {pid,
PID} property is deleted. Now the logs are a lot more readable because
the actual PID is reported as the logging object.