Index details were a waste of time
There's two ways I can track index details, that is, details about
object properties that are pids, whose details are stored in the index:

1. I can go through each object as it's logged, look up each PID, then
look up the ID and icon matching that PID in egre_index

2. I can just leave the top level id and pid properties of the object,
and whatever is doing the logging can track top level IDs and use them
to fill in details of pids in object properties.

The latter means each logger (e.g. DB and web) has to build a map of IDs
to PIDs, and we don't icons. This also means that until the logger has
seen all the PID-ID combinations, it won't have IDs for some PIDs: e.g.
if the first log message is from object A, which references B, then the
ID map won't know about B and the reference to B in A will remain a raw
PID.

e.g. A = [{id, a}, {pid, |1|}, {room, |2|}]
     B = [{id, b}, {pid, |2|}, {player, |1|}]

  When A is logged, the ID->PID map will now have |1| => a,
  but pid |2| won't have a mapped ID.

  So, it'll takes some time for the PID->ID map to "warm up".

The former means we have to do THOUSANDS of Mnesia lookups while we're
logging.

I'm not sure if I care about the icons.

The only reason I care about the Mnesia lookups is that when logic errors
occur in a MUD, Mnesia takes an absolute beating and slows everything
down. This has actually highlighted some areas because Mnesia exerts
significant back-pressure on the graph, making the issue obvious. On the
other hand, faster logging might make it even easier to see that
something is repeating too often.

I'm debating what to do with the Mnesia property storage. I'm thinking I
might have a disk-backed Mnesia table for each object. However, that
would probably be overkill with all the protections that Mnesia has for
multiple processes accessing it. I might just have a simple process that
updates BERT (binary erlang term storage) files with current object
properties. The problem I'm trying to solve is reconstituting objects
from disk if they crash.