DB ID column conflicts with object ID property
I've switched over to using the 'pid' and 'id' properties of each object
in the database. Since the column names are set up to match the object
property keys (for convenience), the object 'id' property collided with
the auto-incrementing primary key column "id" of the database. We opted
for "pk" as the primary key column name because we couldn't think of
anything better and don't really care. Changing the DB column name is
way easier than changing the name of the object 'id' property in
egre_object, ... but we could have easily made an exception in the 1:1
column property key mapping. <shrug>

I've stopped looking up all the pid values of the object properties in
the index Mnesia table to get their corresponding ID and icon. I'm now
just doing that in the database "pid_id" table and the JS procIds map.