WIP: display chains of events
For debugging, I want to see:
- what chain each event is part of
- what events aren't part of a chain, i.e. "dangling"

In order to do this I plan to normalize the events from the code and show
which events spawn other events, and then lay them out in a hierarchy.

Here's a sample so of a map that I put together based on a succeed
function in a module:

  #{module => <<"effect_attack">>,
    matches => #{<<"AttackSource">> => 4,
                 <<"Character">> => 1,
                 <<"EffectType">> => 2,
                 <<"Target">> => 3},
    context => <<"_">>,
    event =>           "1,affect,2,with,3",
    new_event_tuple => "1,roll,2,for,hit,with,2,on,3,with,attack_source,4,effect,5",
    stage => succeed,
    new_event_arg => {var,{64,33},'Event'},
    guard_groups => []}

Here's an idea of what an event my look like with matches from the
custom data object (EGRE_MUD_1 uses a map) and the function guards:

1,roll,2,for,3,with,4,on,5,with,attack_source,6,effect,7 when 1 = hit / effect and 7 = self()

The event might just be indexed on the non-guard portion, i.e. just the
structure of the atom function arguments and variable indexes.