Match guard groups and custom map to event vars
This change "normalizes" the variables in the guard groups and the
custom data map (I'm restricting it to a map for now) so that they match
the normalized, non-atom fields in the event.

Each variable in the event is given an index number. The variables in
the guard groups and the custom map are updated to match these index
numbers.

Example:

Here's the attempt/1 function head and guards:

  attempt({#{owner := Attack}, Props, {Self, affect, Target, with, Attack}, _})
    when Self == self() ->

Here's the current parse transform map that represents that attempt/1
function:

  #{module => <<"effect_attack">>,
    stage => attempt,
    matches => #{<<"Attack">> => 3,<<"Self">> => 1,<<"Target">> => 2},
    context => <<"_">>,
    event => "1,affect,2,with,3",
    custom => [<<"3 == owner">>],
    guard_groups => [[<<"1 == self()">>]]}

Event:          1, affect, 2, with, 3
Custom mapping: 3 must equal the 'owner' field
Guard groups:   1 must equal self()

In this particular cast, this would work out to:
  self(), affect, 2, with, Owner

However, we want to match up other functions that match on
"1,affect,2,with,3" because they'll be in the same chain.