WIP: more work on cataloging the protocol
More fixes in the parse transform to handle all the different code
structures in EGRE_MUD_1.

- Handle attempt/1 functions that can kick off an assortment of
different events, depending on the logic
- Handle guards that are Var1 Op Var2, e.g. Self == Character
- Various different custom map matching situations
- return empty binaries for functions and cases we're ignoring
- Handle another case of attempt/1 that we ignore

It turns out that this protocol cataloguing is not very feasible: I'm
currently at ~700 lines of code. I didn't realize how difficult it would
be to track the events being handled, events being generated, and the
custom map object fields and guard clauses that impose constraints on
events within certain functions. I suppose if I keep up with it, it
won't be as onerous. Hopefully seeing the event chains that form the
implicit protocol will be worth it.