WIP: find children of events
The whole point of the recent parse transform changes is to spot, at
compile time, where event chains have been broken. Events like
{Character, look, Room} are matched by their shape in various different
modules and may spawn other events. If all the shapes of the events
don't "line up" properly, logic isn't triggered correctly and the MUD
doesn't work. It's difficult to remember where all the pieces of an
event chain might be, so a parse transform can find them all and link
them together.

Events in the graph can trigger new events from both the attempt/1 and
succeed/1 functions.

The parse transform code finds the events in the function clauses as
well as new events generated in the function bodies.

This change:

- Serializes all of the event and new_event_tuple values in all the
Event maps, _before_ finding children, so that we can compare new events
that are spawned to events that are handled in attempt/1 and succeed/1.
This is the essence of building event chains: if event 1 is handled and
generates event 2, where is event 2 handled, if at all?

- Creates an "index map" of all the event maps so that event maps can
refer to children as numbers. This way we can store a list of children
that looks like [1, 9, 22]; we can look these up later when we're
serializing the event maps and writing them out to file in hierarchies.

- Fixes a bug where function calls in events weren't updating the
"non-atom value position index".
e.g. {Character, look, self()} might become {1,look,1} instead of
{1,look,2}

TODO:
- Look through all the maps after the children have been added and find
the top-level parents. Top-level parents are not the children of any
other event.

- Print out the events with their children, depth first: print a top
level parent, then for each spawned event print it's children, etc.
Depth first means we don't have all the "sibling" events clumped up
first and then all the children clumped up after.

Here's an example of an event tree:

Top-level Parent
    Spawned Event 1
        Event 1 attempt/1 clause in Module A
        Event 1 succeed/1 clause in Module A
        Event 1 attempt/1 clause in Module B
            Spawned Event 3
    Spawned Event 2
        Event 2 attempt/1 clause in Module C

The "event chains: in this case would be:
1. Top-level parent -> event 1 -> event 3
2. Top-level parent -> event 2