WIP: More progress towards writing event chains
I'm making sure all the fields in each event map are serialized,
figuring out which events have parents by checking which events are
listed as children, and then recursively writing out events, the events
they spawn, and then the events that match those spawn: event -> spawn
-> children.

It looks pretty borken, but it's getting there.

Fixed two bugs:

1: Don't serialize the "Matches" map before using it to turn guard
variables into Match indexes.
e.g.

  Event = {1, buy, 2},
  Matches = #{<<"Character">> => 1,
              <<"Item">> => 2},
  Guard = when Character = self().

That Guard needs to become "when 1 = self()" using the Matches map.
If I serialize the map before the guards code has a chance to use it,
then ... things break.

2. Don't overwrite the Matches map when turning a new event that is
being generated (e.g. "NewEvent = {Self, buy, Item}"), since those
Matches are not relevant to the function head. That is, we need the
matches for the guard clauses and the custom map matching; we could
merge in the "NewEvent" matches, but we don't need them for anything.