WIP: protocol chain discovery
Infer types of spawned events based on "known" custom map fields and
guard clauses. If we see a mapping from a field in the custom map that's
known to be a process, e.g. owner, then we can infer that any variable
'owner' is matched with is a process. With guards, if there's a guard
class that calls is_pid(Var), or is_binary(Var), then we know what the
type of that Var must have been. If we gather these up, then we can
matche the {Var, Type} tuples with the #{Var => Index} map and create a
list of {Type, Index} tuples. Once we have these, we can compare the
{Type, Index} tuples of a spawned event with potential "child" event and
see if there are any conflicts.

e.g. if an attempt/1 clause "spawns" a {1,do,2} event with inferred
types [{bin, 2}], then we know if can't match a potential child event
{1,do,2} if the types are [{pid, 2}]; the types conflict. This should
prevent type-agnostic events from cycling, which we were seeing with
rules_body_part_inject_self.

I'm currently calling events that are generated in an attempt/1 or
succeed/1 function a "spawned" event. Anything event that is triggered
by a spawned event is a child event.

e.g.    <attempt/1 or succeed/1 event>
          <spawned event>
            <child attempt/1 or succeed/1 event>

I'd already started working searching non-attempt/1, non-succeed/1
functions for any events they might generate, so there's a little bit of
extra code in here.

Co-authored-by: Terry Wilcox <terry.wilcox@mac.com>