Test parse transforms
I have a parse transform that I'm working on to detect event chains
based on handled and generated events in the code.

I need unit tests to make sure that different cases are handled.

The gist of it is: compile an "_in" file with the AST translate parse
transform module and write the result to a file. Then compile the "_out"
file with an "id" transform (i.e. "identity", or "leave it the same")
and write that to file. The result of the AST translate parse transform
should equal the "identity transform" of the _out file.

I know what the output of the AST translate parse transformer should
look like as code, but it's really hard to figure out of the AST is
correct. This way, I write the expected code as Erlang, convert it to
the same output format as egre_protocol_ast_translate.erl, and make sure
they match.

Files:

The ./t file is the test runner.

egre_protocol_SUITE.erl is the Common Test test suite.

egre_protocol_SUITE_data is the "data_dir" for the Common Test test
suite, and is returned by proplists:get_value(data_dir, Config).

level_1_call_no_args_in.erl is an Erlang module with a function call
that should be inlined.

level_1_call_no_args_out.erl is an Erlang module that represents what
the _in.erl parse transform forms should look like after
egre_protocol_ast_translate.erl is finished with it.

egre_protocol_id_transform.erl converts parse transform AST forms to
match the output of egre_protocol_ast_translate.erl, but doesn't do any
inlining.