Case expressions for >1 args need data structure
Update the expected code to match what the AST translate transform
should spit out.

If we have a function that takes more than one arguments, we need to
wrap that in a data structure to make it a case statement expression;
further, if we wrap the function arguments in a data structure in the
case expression, then we also have to wrap the matched arguments of the
function clause in a data structure.

e.g. attempt(_) ->
         Call_one_arg =
             case {ok} of  <-- wrap 'ok' in tuple
                 {ok} ->   <-- wrap 'ok' in tuple
                     {return, ok}
             end.
This example, even though there's only one argument, demonstrates
wrapping both the calling arguments (i.e. the case expression) in a
tuple, and then the function clause matched arguments (i.e. the case
clause match pattern) in a tuple