Parse transform warnings are expected
When I'm inlining functions, I'll have code with functions and transform
that using the AST translate transform; then I write the code with
inlined functions, and generate that AST with the "ID" transform.
Then I compare the two, and the code with functions should have been
transformed so it looks like the code with inlined functions.

It's fine if the inlined version of the code has unused variables.
The inlined code is just a code representation of the AST that I'm
expecting, and is only used to search for artifacts within the AST.

Example:

Code that needs to be inlined by the AST translate transformer:
  ░ -module(level_1_call_1_literal_arg_in).
  ░
  ░ -export([attempt/1]).
  ░
  ░ attempt(_) ->
  ░     call_one_arg(ok). <-- function name becomes var
  ░
  ░ call_one_arg(Var) ->
  ░     {return, Var}.

Code that I wrote manually that's already inlined:

  ░ -module(level_1_call_1_literal_arg_out).
  ░
  ░ -export([attempt/1]).
  ░
  ░ attempt(_) ->
  ░     Call_one_arg =   <-- unused var is okay
  ░         case {ok} of
  ░             {ok} ->
  ░                 {return, ok}
  ░         end.