Don't replace var names with literal values
Literal calling arguments should not replace variable names in the
function clause with the literal value.

e.g.

a() ->
    b(ok).

b(Var) ->
    Var.

In this case, when we inline it, the 'ok' should replace the entire var
"Var", not become it's name.

In the AST we want the case clause match to be {tuple, [{atom, ok}]},
not {var, {atom, ok}}, which isn't even legal.

So, we replace vars with literal values, or we replace var names with
var names.

e.g.

a() ->
    B = ok,
    b(B).

b(Var) ->
    Var.

In this case we want to replace the name of the var "Var" with "B", so
that when we inline the b/1 function as a case statement, it references
the logic from the call site scope.