Don't need to see matching parse transforms
I only want to see cases where translated code AST doesn't match the
"identity" AST created from code that matches the intent of the
translation.

It just so happens that it's much easier to compare short ASTs when they
are on a single line and right on top of each other.