Fix bugs ...
- The accumulator returned from rename_form_args/2 is a tuple of the new
body and the RenamedArgMap. We need to capture just the body.

- The match expression for a case clause representing a function clause
needs to wrap the function match in a tuple:

  Where a function can be a(B, C ,D), with parentheses, a case needs to
  be {B, C, D} because it's not wrapped in parameters.

  e.g.

  a(B, C, D) ->
      ok.

  case {B, C, D} of
      {B, C, D} ->        <-- This is the matching function clause
          ok

  Said another way:

      Eshell V15.0 (press Ctrl+G to abort, type help(). for help)
      1> case 1, 2 of
            1, 2 ->
               ok
         end.
      * 1:7: syntax error before: ','

      Eshell V15.0 (press Ctrl+G to abort, type help(). for help)
      1> case {1, 2} of
             {1, 2} ->
                 ok
         end.
      ok

- In rename_form_args/2 we need the {var, Var} function clause to
come before the "is_tuple(Form)" clause so it doesn't get swallowed
up and treated like any other tuple. We have to capture it first in
order to replace the function variable with either a variable or value
from the parent scope.

- When renaming form arguments in a list, make sure not to unwrap the
list by adding it to the accumulator "raw".

  e.g. if the list in {tuple, [{atom, ok}]} was left raw or "unwrapped",
       we'd get {tuple, {atom, ok}}. A tuples elements must be wrapped
       in a list in the AST.

I'm also leaving in the egre_dbg:add/2 call as a convenience so I
remember it's there and don't have to waste a couple brain cycles
figuring it out again.