Add new test, fix major bug
I made a dumb mistake: in order to model a function clause as a case
statement, you only need to change the call to a case:

e.g.

  ░ a() ->
  ░     X = 1,
  ░     b(X).
  ░
  ░ B(Y) when Y == 1 ->
  ░     foo;
  ░ B(Y) ->
  ░     Y.

The inlined case statement would look like this:

  ░ a() ->
  ░     X = 1,
  ░     case {X} of
  ░         {Y} when Y == 1 ->      <-- Y doesn't need to be X
  ░             foo;
  ░         {Y} ->
  ░             Y
  ░     end.

The only thing that has to change with the function clauses is that they
need to be wrapped in a tuple: we matching on the case expression so we
can use the origin variables the function used.

I think it might actually not make any difference whether we change the
function parameter variables to match the parent scope of not.

I think the code would function the same if it was this:

  ░ a() ->
  ░     X = 1,
  ░     case {X} of
  ░         {X} when X == 1 ->
  ░             foo;
  ░         {X} ->
  ░             X
  ░     end.

Now, whether that holds in *every* case, even complicated ones, is a
different question.

The whole point of this exercise is to inline functions so that code in
a future module doesn't have to jump around trying to figure out if an
event handler spawns events through some chain of utility functions.