The solution was actually simple
By which I mean the solution to inline functions in Erlang AST.

I was over-thinking this. When we turn a function call into a case
statement, we don't need to match that case statement to a variable, and
then replace the function call with a variable.

e.g. instead of

attempt(_) ->
    f1(f2(1)).
f1(Y) -> Y + 1.
f2(Z) -> Z * 2.

... becoming

attempt(_) ->
    F2 = case {1} of {Z} -> Z * 2 end,
    F1 = case {F2} of {Y} -> Y + 1 end,
    F1.

... we can just do this:

attempt(_) ->
    case {
        case {1} of
            {Z} ->
                Z * 2
        end
         } of
        {Y} ->
            Y + 1
    end.

... That is, instead of saving function case statement results in a
variable, we can just replace the function calls directly with case
statements in place: there's nothing illegal about having a case
statement being the expression for another case statement.

"Replacing in place" drastically simplified the code. I no longer have
to keep track of whether I'm in a data structure so I know how far back
up the stack I need to pass match-cases.

The only slight twist was that case expressions need to be wrapped in a
data structure, so function clause arguments also have to be wrapped in
a data structure to match. Functions allow multiple expressions, but
case statements don't.