Can't recurse in syntax, LCs not handled
When inlining functions, we never need to inline the code in the body of
a function more than once: if the function calls itself, that's a
runtime phenomenon, not a compile time phnenomenom.

Example:

a() ->
  b().

b() ->
  b().

  We can inline this as:

a() ->
  case {} of     % All three of these lines contain the code
    {} ->        % for b(); it's irellevant that the code for
      b()        % b() contains a call to itself.
  end.

The problem we're trying to solve is finding out action-reaction event
pairs in the code. Inlining the functions makes it so we don't have to
search multiple functions to find reaction events, i.e. events that are
generated while handling an action event. See the tests for another
example.

LCs, list comprehensions, were not handled in the parse transform. List
comprehensions can contain calls to utlity functions that generate
events. See tests for an example.