WIP: continue implementing event parsing with TDD
Move parse transform output samples into a .hrl file.
I'm still a bit hazy on *why* Erlang can find a .hrl file in the test
data directory (..._SUITE_data/) but it does find it.
We're going to have a ton of sample data representing outputs from the
parse transform and I don't want them cluttering up the code.

So far we have a test for an API function that neither handles nor
generates events (which was easy) and a test for a "terminal" event
where an API function handles an event, but doesn't generate a reaction
event.

We need the type inference data to make sure events that match on shape,
but not type, don't recurse:

  {attack, <<"bob">>}         -> {attack, <0.1.0>}
                           vs.
  {attack, 1}, [{1, binary}]  -> {attack, 1}, [{1, pid}]

  The events without type data on the top line would match and recurse;
  however, the events on the bottom line won't match because, although,
  they've had their shape normalied with variable index placeholders,
  their types don't match. (binary /= pid).

Since we're looking for action-reaction event pairs, for a terminal
event we're just leaving the reaction event as undefined for each of its
three attributes: indexed event, indexed variables, and indexed type
inference. There are currently no cases in code which uses EGRE where an
API function (attempt/1, succeed/1) _doesn't_ handle a specific event,
but _does_ generate an event ... I don't think. This could happen if we
were to generate metrics though; I suppose we may have to handle that
case in the future. In that case, the action event fields could also
just be undefined: in the final protocol output these might not be
matched to any parents.

Example:

BERT file record for metrics reaction event:

{<<"some_rules_module">>, attempt,
 undefined, undefined, undefined,       %% action event
 {metrics, 1, inc}, [{1, pid}], [{1, Bar}]  %% reaction event
}

Sample event chain with metrics event matching everything:

{1, move, 2, from, 3} ->
  {metrics, 1, inc}
  {1, move, to, 2, from, 3} ->
    {metrics, 1, inc}
    {send, 1, 2} ->
      {metrics, 1, inc}